+++ core/trunk/src/main/java/net/indrix/arara/servlets/sound/DeleteSoundServlet.java
@@ -34,6 +34,7 @@ import org.apache.log4j.Logger;
  * To change the template for this generated type comment go to
  * Window>Preferences>Java>Code Generation>Code and Comments
  */
+@SuppressWarnings("serial")
 public class DeleteSoundServlet extends HttpServlet {
     /**
      * Logger object
@@ -50,34 +51,57 @@ public class DeleteSoundServlet extends HttpServlet {
         String nextPage = null;
         List <String>errors = new ArrayList<String>();
         String soundId = req.getParameter("soundId");
+        String ownerId = req.getParameter("userId");
+        
+        HttpSession session = req.getSession();
+        User user = (User) session.getAttribute(ServletConstants.USER_KEY);
+        
+        if (user == null) {
+            logger.debug("User null.");
+            errors.add(ServletConstants.USER_NOT_LOGGED);
+            // put errors in request
+            req.setAttribute(ServletConstants.ERRORS_KEY, errors);
+            nextPage = ServletConstants.LOGIN_PAGE;
+        } else {
+            int photoAuthorId = (ownerId != null) ? Integer.parseInt(ownerId) : -1;
+            if (user.getId() == photoAuthorId){
+                SoundModel model = new SoundModel();
+                try {
+                    int id = Integer.parseInt(soundId);
+                    model.delete(id);
+                    logger.debug("Sound deleted..." + soundId);
 
-        SoundModel model = new SoundModel();
-        try {
-            int id = Integer.parseInt(soundId);
-            model.delete(id);
-            logger.debug("Sound deleted..." + soundId);
+                    loggerActions.info("User " + user.getLogin() + " from IP " + req.getRemoteAddr() + " has deleted the sound " + soundId);
 
-            HttpSession session = req.getSession();
-            User user = (User) session.getAttribute(ServletConstants.USER_KEY);           
-            loggerActions.info("User " + user.getLogin() + " from IP " + req.getRemoteAddr() + " has deleted the sound " + soundId);
-
-            nextPage = ServletConstants.FRAME_PAGE;
-            String pageToShow = ServletConstants.DELETED_SOUND_PAGE;
-            
-            req.setAttribute(ServletConstants.NEXT_PAGE_KEY, nextPage);
-            req.setAttribute(ServletConstants.PAGE_TO_SHOW_KEY, pageToShow);          
-        } catch (DatabaseDownException e) {
-            logger.debug("DatabaseDownException.....");
-            errors.add(ServletConstants.DATABASE_ERROR);
-        } catch (SQLException e) {
-            logger.debug("SQLException.....", e);
-            errors.add(ServletConstants.DATABASE_ERROR);
+                    nextPage = ServletConstants.FRAME_PAGE;
+                    String pageToShow = ServletConstants.DELETED_SOUND_PAGE;
+                    
+                    req.setAttribute(ServletConstants.NEXT_PAGE_KEY, nextPage);
+                    req.setAttribute(ServletConstants.PAGE_TO_SHOW_KEY, pageToShow);          
+                } catch (DatabaseDownException e) {
+                    logger.debug("DatabaseDownException.....");
+                    errors.add(ServletConstants.DATABASE_ERROR);
+                    nextPage = ServletConstants.FRAME_PAGE;
+                } catch (SQLException e) {
+                    logger.debug("SQLException.....", e);
+                    errors.add(ServletConstants.DATABASE_ERROR);
+                    nextPage = ServletConstants.FRAME_PAGE;
+                }
+                
+            } else {
+                // user do not have the right to delete the photo
+                nextPage = ServletConstants.FRAME_PAGE;
+                String pageToShow = ServletConstants.SHOW_MESSAGE_PAGE;
+                String messageKey = "operation.denied";
+                req.setAttribute(ServletConstants.SEND_EMAIL_MESSAGE_KEY, messageKey);
+                req.setAttribute(ServletConstants.PAGE_TO_SHOW_KEY, pageToShow);                
+            }
         }
+
         if (!errors.isEmpty()) {
             logger.debug("errors is not null.");
             // put errors in request
-            req.setAttribute(ServletConstants.ERRORS_KEY, errors);
-            nextPage = ServletConstants.UPLOAD_PAGE;
+            req.setAttribute(ServletConstants.ERRORS_KEY, errors);            
         }
         
         dispatcher = context.getRequestDispatcher(nextPage);
